# docs/context/api/openapi.yaml
openapi: 3.1.0
info:
  title: Paper Engineering Assistant API
  version: 0.1.0
  description: |
    Canonical REST contract for backend routes used by desktop literature and governance workflows.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /literature/retrieve:
    post:
      operationId: retrieveLiterature
      summary: Retrieve relevant literature chunks from active embedding versions.
      tags: [literature]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiteratureRetrieveRequest'
      responses:
        '200':
          description: Retrieval result with chunk-level evidence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiteratureRetrieveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /paper-projects/{id}/literature-links/from-topic:
    post:
      operationId: syncPaperLiteratureFromTopic
      summary: Sync in-scope topic literature into paper links.
      tags: [literature]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncPaperLiteratureFromTopicRequest'
      responses:
        '200':
          description: Sync summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPaperLiteratureFromTopicResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /paper-projects/{id}/literature:
    get:
      operationId: getPaperLiterature
      summary: Get literature links under a paper.
      tags: [literature]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Paper literature links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPaperLiteratureResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /paper-projects/{id}/literature-links/{linkId}:
    patch:
      operationId: updatePaperLiteratureLink
      summary: Update citation status or note of a paper literature link.
      tags: [literature]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: linkId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePaperLiteratureLinkRequest'
      responses:
        '200':
          description: Updated paper literature link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePaperLiteratureLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /literature/{literatureId}/metadata:
    get:
      operationId: getLiteratureMetadata
      summary: Get editable metadata fields of one literature.
      tags: [literature]
      parameters:
        - name: literatureId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Literature metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLiteratureMetadataResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateLiteratureMetadata
      summary: Update literature metadata fields.
      tags: [literature]
      parameters:
        - name: literatureId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLiteratureMetadataRequest'
      responses:
        '200':
          description: Updated literature metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateLiteratureMetadataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  responses:
    BadRequest:
      description: Invalid request payload or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Business conflict or version conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          additionalProperties: false
      additionalProperties: false
...
    RightsClass:
      type: string
      enum: [OA, USER_AUTH, RESTRICTED, UNKNOWN]
...
    PaperCitationStatus:
      type: string
      enum: [seeded, selected, used, cited, dropped]
...
    SyncPaperLiteratureFromTopicRequest:
      type: object
      required: [topic_id]
      properties:
        topic_id:
          type: string
          minLength: 1
      additionalProperties: false
    SyncPaperLiteratureFromTopicResponse:
      type: object
      required: [paper_id, topic_id, linked_count, skipped_count]
      properties:
        paper_id:
          type: string
        topic_id:
          type: string
        linked_count:
          type: integer
        skipped_count:
          type: integer
      additionalProperties: false
    PaperLiteratureLinkView:
      type: object
      required:
        - link_id
        - paper_id
        - topic_id
        - literature_id
        - citation_status
        - note
        - created_at
        - updated_at
        - title
        - authors
        - year
        - doi
        - arxiv_id
        - source_provider
        - source_url
        - tags
      properties:
        link_id:
          type: string
        paper_id:
          type: string
        topic_id:
          type: [string, 'null']
        literature_id:
          type: string
        citation_status:
          $ref: '#/components/schemas/PaperCitationStatus'
        note:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: [integer, 'null']
        doi:
          type: [string, 'null']
        arxiv_id:
          type: [string, 'null']
        source_provider:
          type: [string, 'null']
        source_url:
          type: [string, 'null']
        tags:
          type: array
          items:
            type: string
      additionalProperties: false
    GetPaperLiteratureResponse:
      type: object
      required: [paper_id, items]
      properties:
        paper_id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaperLiteratureLinkView'
      additionalProperties: false
    UpdatePaperLiteratureLinkRequest:
      type: object
      properties:
        citation_status:
          $ref: '#/components/schemas/PaperCitationStatus'
        note:
          type: string
      additionalProperties: false
      minProperties: 1
    UpdatePaperLiteratureLinkResponse:
      type: object
      required: [paper_id, item]
      properties:
        paper_id:
          type: string
        item:
          $ref: '#/components/schemas/PaperLiteratureLinkView'
      additionalProperties: false
...
    GetLiteratureMetadataResponse:
      type: object
      required:
        - literature_id
        - title
        - abstract
        - key_content_digest
        - updated_at
      properties:
        literature_id:
          type: string
        title:
          type: string
        abstract:
          type: [string, 'null']
        key_content_digest:
          type: [string, 'null']
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    LiteratureRetrieveRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        topic_id:
          type: string
          minLength: 1
        paper_id:
          type: string
          minLength: 1
        top_k:
          type: integer
          minimum: 1
          maximum: 30
          default: 10
        evidence_per_literature:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
      additionalProperties: false
    LiteratureRetrieveEvidenceChunk:
      type: object
      required:
        - chunk_id
        - text
        - start_offset
        - end_offset
        - hybrid_score
        - vector_score
        - lexical_score
      properties:
        chunk_id:
          type: string
        text:
          type: string
        start_offset:
          type: integer
        end_offset:
          type: integer
        hybrid_score:
          type: number
          minimum: 0
          maximum: 1
        vector_score:
          type: number
          minimum: 0
          maximum: 1
        lexical_score:
          type: number
          minimum: 0
          maximum: 1
      additionalProperties: false
    LiteratureRetrieveHit:
      type: object
      required:
        - literature_id
        - title
        - embedding_version_id
        - hybrid_score
        - vector_score
        - lexical_score
        - evidence_chunks
      properties:
        literature_id:
          type: string
        title:
          type: string
        embedding_version_id:
          type: string
        hybrid_score:
          type: number
          minimum: 0
          maximum: 1
        vector_score:
          type: number
          minimum: 0
          maximum: 1
        lexical_score:
          type: number
          minimum: 0
          maximum: 1
        evidence_chunks:
          type: array
          items:
            $ref: '#/components/schemas/LiteratureRetrieveEvidenceChunk'
      additionalProperties: false
    LiteratureRetrieveResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LiteratureRetrieveHit'
        meta:
          type: object
          required: [query_tokens, profiles_used, skipped_profiles]
          properties:
            query_tokens:
              type: array
              items:
                type: string
            profiles_used:
              type: array
              items:
                type: object
                required: [provider, model, dimension, literature_count]
                properties:
                  provider:
                    type: string
                  model:
                    type: string
                  dimension:
                    type: integer
                  literature_count:
                    type: integer
                additionalProperties: false
            skipped_profiles:
              type: array
              items:
                type: object
                required: [provider, model, dimension, reason]
                properties:
                  provider:
                    type: string
                  model:
                    type: string
                  dimension:
                    type: integer
                  reason:
                    type: string
                additionalProperties: false
          additionalProperties: false
      additionalProperties: false
    UpdateLiteratureMetadataRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        abstract:
          type: [string, 'null']
        key_content_digest:
          type: [string, 'null']
        authors:
          type: array
          items:
            type: string
            minLength: 1
        year:
          type: [integer, 'null']
          minimum: 1900
          maximum: 2100
        doi:
          type: [string, 'null']
        arxiv_id:
          type: [string, 'null']
        rights_class:
          $ref: '#/components/schemas/RightsClass'
        tags:
          type: array
          items:
            type: string
            minLength: 1
      additionalProperties: false
      minProperties: 1
    UpdateLiteratureMetadataResponse:
      type: object
      required:
        - literature_id
        - title
        - abstract
        - key_content_digest
        - authors
        - year
        - doi
        - arxiv_id
        - rights_class
        - tags
        - updated_at
      properties:
        literature_id:
          type: string
        title:
          type: string
        abstract:
          type: [string, 'null']
        key_content_digest:
          type: [string, 'null']
        authors:
          type: array
          items:
            type: string
        year:
          type: [integer, 'null']
        doi:
          type: [string, 'null']
        arxiv_id:
          type: [string, 'null']
        rights_class:
          $ref: '#/components/schemas/RightsClass'
        tags:
          type: array
          items:
            type: string
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
...